% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stickyFigure.R
\name{draw.stick}
\alias{draw.stick}
\title{Plot stick figures}
\usage{
draw.stick(x, y, scale = 1, arms = "down", gender = "male", lwd = 3,
  clcol = "white", face = "happy", linecol = gray(0.3), hat = NA)
}
\arguments{
\item{x}{integer}
}
\description{
Plot stick figures
}
\seealso{
\url{https://github.com/EconometricsBySimulation/R-Graphics/blob/master/Stick-Figures/draw.stick.R}
}

