% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic.R
\name{input_classic}
\alias{input_classic}
\title{Classic R approached to data analysis}
\usage{
input_classic(pdata_file, exprs_file)
}
\arguments{
\item{pdata_file}{character(1) The path to the phenotype data file}

\item{exprs_file}{character(1) The path to the expression data file}
}
\value{
A data.frame() containing samples as rows and phenotype data and 
expression data as columns.
}
\description{
The functions here illustrate how an 'old school' R usere might approach data
management analysis.
}
\details{
There are two files. They are from a microarray experiment. The 
first file, ALL-phenoData.csv describes the samples. The second file, 
ALL-expression.csv are the normalized experssion values for 12k probsets 
across 128 samples. The files originally came form sheets in excel, and were
exported from Excel using export-to-csv
}

