% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi.R
\name{hi}
\alias{hi}
\title{Help start conversations}
\usage{
hi(who)
}
\arguments{
\item{who}{character(1) The name of a person you wish to start a 
conversation with}
}
\value{
character(1) Line of text to be used when starting conversations
}
\description{
This function generates some helpful text that can be used to start
conversations in all kinds of awkward social situtations.
}
\examples{
hi("Lori Shepherd")

}

